Fix phased sync initialization
Turns out ApplicationInitializedListener is too early to interact with
the registry flags.

Had to add the phased sync init in various places in test setup.
I intended to use TestEnvironmentRule initially as well, but that's too
early for tests too, so modified FixtureRule implementations and then
also the older test classes that don't use the rules.

Bug: 384022658
Test: N/A
Change-Id: I2039561dac94aa2111b8b3d43d431821b9b94d7b