Fix maven action analysis-on-write error
The gmaven import action already has an override to allow it to do
analysis on the EDT, but was missing one to do it within a write action.
Fixing that.

This was not caught by tests, even though we have tests covering this
scenario. That happened because the tests were invoking the action
directly, and were not entering a write action first (like the product
actually does). By switching the tests to use a fixture utility method
for invocation, we now have tests that appropriately catch this case.

Fixes: 416425405
Test: AndroidMavenImportIntentionActionTest
Change-Id: I891a7d8265d8739e6d438992d970bfde76066b23