Introduce some convenience methods
Provide hasRegisteredDependency for the common case of checking
whether a dependency exists or not, without being concerned about any
additional details of that dependency.

Provide registerDependency for a WellKnownMavenArtifactId argument,
delegating directly to the concrete method with a specialized
dependency id.

Adjust some callers in production to use the convenience methods.

Bug: 279149611
Test: added
Change-Id: Ifd5e97d6716985dde7ac85a907a7e48245d4e340