Clean up icon-related code in AVD and profilers Compose code
This CL is cleaning up a few unnecessary parameters passed to the Icon
composable: the iconClass is derived by default from the iconKey, so it
should not be necessary to pass it explicitly.

Note that this does not apply to *IconActionButtons, which are currently
suffering from a bug tracked by JEWEL-817.

Bug: b/409420047
Test: no functional changes
Change-Id: I12bd12c9b2836727fa2205fac5237cd7ad755c6a