Move RenderErrorTest to flows
So we wait for the right result. After moving to Truth, the failures
showed that some times all the issues do not arrive at the same time so
the test will fail if the issues arrive in the wrong order.
This changes the test so it waits for 20s for the right issue to arrive.

Test: RenderErrorTest
Bug: N/A
Change-Id: Iae4c6ae9d2e25c98162dfaf668e7f75756dbc2a8