Change printing of AgpVersionSoftwareEnvironmentDescriptor
Include the JDK version (and the compileSdk) in the string
representation, so that the test that verifies that .selected returns
a single value doesn't print a confusing error message when that isn't
true.

Generate a string without spaces so that fusion is less likely to
report test names like "32, m=V1)]".

Bug: none filed
Test: test-only change
Change-Id: I778c262d71171a2ee0730b70f66506c8f150b754