Keep moving project state management
from the `QuerySyncProject` to the `QuerySyncManager` class. The former
will soon only hold the project definition and a collection of services
configured according to the project definition.

Inline `sync` method so that `updateProjectSnapshot` calls end up in the
manager so that they can be later moved to the `runOperation` level.

The enable code analysis methods will be updated in upcoming changes.

Bug: 417935665
Test: n/a
Change-Id: I0efa5853d6c5cb6c6df4e439ba37fe4be674d7aa