Fix ScenePicker concurrency
This makes two changes that allow the ScenePicker to work with the new
async model.
It introduces a "double-buffer" where one ScenePicker is used to add
shapes and one to read. This allows for the new one to be build in the
background while the "old" one is still being used to resolve mouse
clicks.
Additionally we move the SecondarySelector to be an immutable data class
that does not allow for changes. This ensures that they are usable from
multiple threads and cannot be disposed in the middle of a rendering.

Test: Existing
Fixes: 410750497
Change-Id: I936b2ad48f5fdcb4839720f11d9fe4e4c2c6de2c