multi_info_file mode is the only supported one
Delete all test data targets generated using the old mode. Some of the
tests using them will be replaces by `.bzl` tests and some unfortunately
are lost although their value is questionable considering that they rely
on `.java-info.txt` files different from those used in the production.

Moreover, most of these tests are making assertions on
ProjectProto.Project structure, which is changing and it is not feasible
to maintain these tests in this form.

Bug: n/a
Test: n/a
Change-Id: I0720b46ac805c944488ffb48e3b81f748c59040f