Use runBlocking from non-cancellable context
The platform implementation (nonblockingResolveGradleJvmInfo) uses
runBlockingCancellable, which asserts we're in a cancellable context,
but in this view context, there is no cancellable job. Example failures:
- java.lang.IllegalStateException: This method is forbidden on EDT
because it does not pump the event queue. Switch to a BGT, or use
 com.intellij.openapi.progress.TasksKt.runWithModalProgressBlocking.
- java.lang.IllegalStateException: There is no ProgressIndicator or Job
 in this thread, the current job is not cancellable.

Ideally we should tie the operation to the view's lifecycle, add the
appropriate UI hooks, spinners and all that, but this fix is better than
crashing.

Bug: 415709494
Test: manually tested
Change-Id: I1e3efd6292ab000061698471458f246a95c2b98c