Fix enabling check to download external annotation for dependencies
The dependencies resolution of external annotation was always
marked as disabled for studio, causing user overridden option
via Gradle settings to not being stored.

Initially, this was disabled because was causing several
unnecessary rootChanged events causing significant slowness.

Platform issue got fixed as part of IDEA-323242 and IDEA-322079
so now unnecessary rootChanged events aren't longer triggered.
In addition, this option is disabled by default so users has
to manually change the settings. However, if in the future
we would like to avoid this then will require also to hide
the component from the UI to avoid creating confusion to users.

Bug: 317952220
Fixes: 375228136
Test: Manual
Change-Id: I2f74c4eada84e664f79392b4f6d54b47489f8dcd