Implement a ProjectSystemToken for AndroidMavenImport
Mirror the existing pattern of calls within
AndroidMavenImportIntentionAction, providing token methods to check
for existing (transitive, resolved) dependency, and for registering a
new dependency.

We can use the presence or absence of an applicable token to replace
the checks about whether the module system can register dependencies:
if we have a token, we assume/require that that token can do what we
ask of it (which is reasonable, since this is a specialized
interface).  This resolves a theoretical issue where a project system
might be able to register IMPLEMENTATION dependencies but not, for
example ANNOTATION_PROCESSOR dependencies.  (No such project system
exists currently.)

Implement this token interface for the Gradle project system,
preserving all existing behaviour.

Bug: 279149611
Bug: 279886738
Test: existing
Change-Id: If4a375aa58a11504ef816fa58dab18a22c7ec034