Report unknown WFF versions as warnings only in Studio
Reporting unknown WFF versions as errors can be problematic because
sometimes the version is valid, but we haven't yet added it to AGP/
Studio. In that case, reporting it as an error can be confusing to the
user.

Instead, through this CL, we report it as a warning. This CL also
removes the check from AGP in favor of Studio.

Bug: 406968509
Test: UnknownWFFVersionInspectionTest
Change-Id: Ided0898b0cf3033c67956cb9ba74a5c3c4672d8b