Use predict in assertion
In some cases, we cannot guarantee the output returned by provider.
Instead, we only need to verify it contains something endwith .jar or
it's not an empty string. With this cl, we provide factories for string
and depset files. It will only check prediction withtout checking length
of list and it's exact path/ content.

Bug: b/425384016
Test: existings tests
Change-Id: I551bdd3f1c872a63d3cf0a5ea6104f3bcf98f68f