Update IDE model snapshots caused by KGP upgrade
In the other change in the same topic, AGP will now pull in KGP 2.2.0
to the build script classpath. This means that some test projects that
are using a lower version of KGP will be upgraded to KGP 2.2.0
automatically.

This commit updates the IDE model snapshots that are impacted by that
change.

Bug: 431147146
Test: CurrentAgp.testSimpleApplicationWithScreenshotTest
Change-Id: I04625f4f6eb83c53ac734adc6a82c7b6a0559662