Create reusable "next API" flags for the SDK Upgrade Assistant
Instead of creating new flags and remembering to update the lint
quickfix every year, we can define a pair of reusable flags.

TSDKVUA_API_NEXT defines the current "next" API version and
TSDKVUA_API_NEXT_ENABLE defines whether the "next" API is enabled in the
assistant. Since I'm starting work on the API 36 cycle, I went ahead and
set them to 36 and `false` to match that.

The lint quickfix now uses these two flags to determine the max API
version for suggesting the assistant. It's an inclusive upper bound so
we want the source API of the latest public upgrade path.
TSDKVUA_API_NEXT is the target API of the latest upgrade path and
TSDKVUA_API_NEXT_ENABLED tells us whether that path is public. If it is
then we can just decrement it once to get the version we want. If it is
private we need to decrement it twice. With this, we should no longer
need to update the quickfix because of new APIs.

The flag update cycle going forward will be something like this:
1. Start a new cycle. Move the old "next" tutorial outside of the flag.
2. Increment TSDKVUA_API_NEXT and set TSDKVUA_API_NEXT_ENABLE to `false`
   (or "up to canary" or similar)
3. Add a new "next" tutorial inside the flag.
4. When ready to publish to stable set TSDKVUA_API_NEXT_ENABLE to `true`
   (or "up to stable" or similar)

Fixes: 372494165
Test: N/A
Change-Id: I208edbc79694c4f563596cb07c8e7077b878a8ea