Remove SystemImageStateService
The SDK can change, so we shouldn't store something based on it in an
application-level service. Instead, create the flow directly when
needed. The RepoManager performs caching, so this is not particularly
expensive.

Reimplement the system image flow using the new suspend load methods in
RepoManager. This is simpler and allows loading local and remote in
parallel.

This removes obstacles in ISystemImages to refreshing images after
regaining network connectivity; however, the RemoteRepoLoaderImpl also
caches the failed load, so this does not yet fix b/402102815.

Bug: 402102815
Test: ISystemImagesTest
Change-Id: I26a5c040ae908ebbe7a4a191c2d95f9ab6394b88