Fix ComposeRenderErrorContributorTest not running from IDEA
When running ComposeRenderErrorContributorTest from IntelliJ, we were
crashing because the Application was not being set on time. This was
happening because StudioHtmlLinkManager was being instantiated in the
test constructor and it requires the Application to be able to obtain
the NotificationGroupManager service.

The issue is fixed by moving StudioHtmlLinkManager's instantiation to
a setup method, so the AndroidProjectRule has time to set up the
Application properly before it's needed.

Bug: N/A
Test: ComposeRenderErrorContributorTest
Change-Id: Id75310034c1f02248701b585b452969cee7018f0