Implement TargetPattern on top of Label's
Target patterns is "a special kind of a label" with wildcards allowed at
the end of the build package path or instead of the target name.

Support for target patterns is needed to handle the `Targets:` section
of `.bazelproject` files in query sync, specifically to filter query
results that include all targets withing the project scope to the
matching ones without running another query.

Examples:
  //build/package:target - a simple label is a valid pattern
  //all/packages/under/...
  //all/targets/in:*

  etc.

Bug: 425724950
Test: TargetPatternTest
Change-Id: I3419f475349dcd431a261fbaf6ba6430790a4998