Fix a possibly-latent issue in addDependenciesWithUiConfirmation
Although the type signatures defend us from most problems, in this
instance we effectively have a transformation from
RegisteredDependencyId to RegisteredDependencyId (for example, from a
wildcard-version to a concrete version).  This means that we can lose
the association between an id as computed from the input of this
method, and the id as attempted to be registered -- which in turns
means that we can return the wrong answer, as we fail to map back from
errors to the ids we provided.

Fix this by propagating our original data through, and matching things
up at interface boundaries.  Various methods acquire a generic type
parameter in order to be able to preserve the inputs; fight a bit with
generic signatures until things compile again.

This error was not picked up in tests because the integration-ish
tests involving this use the TestProjectSystem, which itself is set up
to have artifacts named like wildcards, so the transformation applied
from looking up is effectively the identity.  The next CL strengthens
this setup (and thereby exposed the problem.)

Bug: 279886738
Test: existing tests adjusted, next CL
Change-Id: Id077e280cc367cb3a8e5c1fa998edeaaaaf915bf