Do not prefer build api for queries
Switching to a different build invoker just for queries does not solve
any "real" problem. It is safe to assume that the query result (
including a failure) represents the citc snapshot before the query
was initiated and any changes made later will result in the current
query data being invalidated. In the worst case the IDE will re-query
the same data next time.

This change aims to reduce the number of alternative execution modes
in the future.

Bug: n/a
Test: n/a
Change-Id: I4e90fb470651bd8f7317c2d31394c03110fa44d7