Delete workaround for running queries in snapshots
This approach did not work well since bazel does not support it and
simply launches a new bazel daemon when launched in a snapshot
directory.

Instead ASwB will keep running bazel in the workspace root and assume
that the output represents the state of the workspace as it was at the
moment when the query was launched. If anything has changed it will be
detected by other means and the current state will be invalidated.

Bug: n/a
Test: n/a
Change-Id: I81a3848510f6ee525634414b855cc7e9fd2440b2