Remove UnsupportedJdkMinimumVersionIssueChecker when using JDK 7
Originally the check was introduced in Icebbe57d0 which at that time
made completely sense to display quick fixes to users. However, now
after several years, when executing any Gradle task/sync, platform
validates if specified JDK version is supported before proceeding
and resulting in UnsupportedGradleJvmByIdeaException.

This check is implemented on GradleExecutionHelper and uses
GradleJvmSupportMatrix.isJavaSupportedByIdea to determine if specified
JDK is supported, resulting in case of JDK 7 to fail before studio
implementation UnsupportedJdkMinimumVersionIssueChecker is triggered.

Since our implementation was already ignored in the last +1 year with
the introduction of IDEA-342292, there's no need of keeping it.
Separately, the message coming from platform will be improved
adding quickfixes as part of b/407841529.

Bug: 388794527
Test: Existing
Change-Id: I532cd211e5ce16ac1efe446e852de732aaa7192b