Pass query via --query_file when supported
Since we don't know exact limits on the query size in both cases
prefer --query_file, which can support larger queries in all currently
supported cases.

Do not check the size explicitly and let the build invocation fail. If
we keep receiving query too large errors we will make the IDE fall back
to the full query mode (once we learn how failures are reported)

Right now the IDE tries to use --query_file when it is not supported and
when the underlying invoker can handle much bigger query sizes without
it.

Bug: 428180627
Test: n/a (existing)
Change-Id: I43fef94d83299eaa4801f6958baba37fdfae9ae4