Add new render panel based on OnDeviceRendererModel
The new renderer renders the draw instruction from OnDeviceRendererModel
and dispatches user input to it.

This approach to rendering is much simpler since it doesn't need to
handle 3D mode and images.

This is a step in the direction of using OnDeviceRendererModel as the
only render model in Layout Inspector.

Test: NewStudioRendererPanelTest
Bug: NA
Change-Id: Id43a28fab836134604c103912bb35817df260219