avoids access to package-private members
Fixes: 429614973.

This change fixes the cases when a class accesses a package-private
member (or a protected member not from a subclass) which is declared in
a different module. For example, class CpuProfilerContextMenuInstaller
from module 'intellij.android.profilers.ui' calls a method
'CpuProfiler.generateCaptureFileName' which is declared in a module
'intellij.android.profilers' module, and this method had
'package-private' visibility before.

Such code works if both classes are loaded in the same package and
loaded by the same classloader, but will fail with IllegalAccessError if
modules are loaded by different classloaders. This may happen if
modules become registered differently in the plugin.xml file, so it's
better to get rid of such cases to avoid problems in the future.

Also, this allows not including package-private members to ABI JARs to
improve compilation avoidance in Bazel.

All code in intellij monorepo was processed to fix such cases, see
https://youtrack.jetbrains.com/issue/IJPL-149126.

This commit fixes remaining cases in Android plugin source code:
* several methods were made public;
* the return type of MemoryObjectTreeNode.getParent was refined and this
method is used instead of accessing package-private field;
* code in MemoryInstanceDetailsView.repeatedlyExpandFirstReference was
refactored to used public methods instead of package-private field;
* explicit suppression for the inspection which reports such cases
was added in K*AndroidKotlinResourceExternalAnnotator files.

Test: the commit is a refactoring, so no new tests are needed.

Change-Id: Ic884dfc5fe9c6c954a9f0213ce4721db1bd4dced