Fix wrong WFF being used in the XML provider
When the merged manifest wasn't built yet, we would fall back to another
WFF version.

In some cases this version would be wrong and the file would highlight
mistakes due to features from a higher version being used.

Instead, with this CL, we only fall back if we can't find the WFF
version from a valid merged manifest.

Bug: 392849349
Test: updated existing
Change-Id: I701259d832d0bdf99abd40e7dd525df422b063ba