Separate StudioFlags for Query State banner for Chat and Agent Mode
On StudioBot we'll have separate checks for each mode, so we need to
different flags.

Since the banner in Chat is only useful when there's a thinking stream
and we don't have yet a way to know if we expect a thinking stream, the
flag under Chat mode has been set up to DEV channel.

While the Agent Mode status banner is generally more useful regardless
if there's thinking stream since it shows tool usage, so it's enabled
up to STABLE.

Bug: 430563105
Test: n/a
Change-Id: I31ff122dc8c1b2c0c68e2e49c317fccbd5a2e338