Auto Deobfuscate Stack Traces
If a stack trace contains a `pg_map_id`, attempt to find a `mapping.txt`
file in the default location
(`<app-module>/build/outputs/mapping/<variant>/mapping.txt`.

If found, use it deobfuscate the stack trace.

We keep an instance of the `RetraceCommand` in memory for the last
encountered `pg_map_id` in order to improve performance since the first
usage of the retracer is expensive.

Bug: NA
Test: Added
Change-Id: I3d6d9cb1d1eb4605b181a2cd7518353c2c69cf02