Release SplitEditorToolbar reference when disposing editor
In SeamlessTextEditorWithPreview, we keep a reference to a
SplitEditorToolbar in order to be able to hide and show it when the
editor content changes (e.g. when Composable imports are added/removed).

This reference, however, was leaking after the editor was diposed, so we
need to clear the toolbarComponent property when disposing the editor.

Fixes: 425753989
Test: SourceCodeEditorWithMultiRepresentationPreviewTest
Change-Id: I51086bbe82a336d3c6b15d10f24f318c64c83704