Fix debugger "View bitmap" functionality
The View bitmap action in the debugger was not visible in the recent
versions of Android Studio. It seems that a platform change cause the
new getFullValueEvaluator method to be called while it was not
implemented in the BitmapRenderer.
This adds the implementation for the new method and restores the action.

Fixes: 420956733
Test: Manually verified
Change-Id: I34bf9daa60d04548bffcddc5cec77d45b7af7530