Avoid loading classes in VerifyGradleTokensAction
With IntelliJ 2025.2 the MultiRoutingFileSystem was marked as the
default to create a more powerful and unified VFS and now
FileSystems.getDefault() no longer returns LinuxFileSystem.

That change cause io.github.classgraph.ClassGraph library to fail
when converting classpath path as string into a file with
UnsupportedOperationException: Path not associated with default
file system.

Current implementation uses ClassInfo.loadClass() to load the class
first and check if implements certain interfaces via isAssignableFrom().
To archive the same behaviour, reduce class loading overhead and
bypass the library issue with latest version of platform, the
ClassInfo.implementsInterface() method will be used instead without
the need of loading the class and archiving 5x performance gain.

Bug: 429413519
Test: GradleTokensTest
Change-Id: Ibcdd9af90300f672d5bb41fa4b0aa2fd0c800692