Allow creating RenderResult future in non-test environment
We need to allow rendering previews in other parts of Android Studio,
outside our regular render pipeline (within UI Preview tools). We can
achieve that by leveraging the createRenderTaskFuture (now) test-only
function. Since we don't want new clients to be responsible for manually
disposing the RenderTask, we keep this function visible only for tests,
but reuse it into a newly added createRenderResultFuture, which creates
the RenderTask future and hooks up its disposal in the RenderResult
future that is created and returned by the function. This new function
can be used in non-test environment without the worry of making the
callers responsible for disposing the render task.

Bug: 422127061
Test: updated affected tests
Change-Id: I3a09401291792104e55b67fb4617a43802dce075