Unify some test utilities
Make the almost-identical code in TestConfigurationTesting and
AndroidGradleTestConfigurationTestingUtil manifestly identical,
leaving only some convenience static functions in those two classes,
and moving all the rest to a new utility class.  This is in
preparation to making these utilities aware of read-locks in order to
be able to port tests using them to JUnit4.

Bug: none filed
Test: existing
Change-Id: I25f2ae992bd97bff962d6fa535e789b586716a2a