Extract ToggleOverlayAction from toolbar
This change is in preparation of adding support for overlays to
on-device rendering.

ToggleOverlayAction is now provided from the users of the toolbar,
this way we can be more flexible on how to set and get the image.

Also instead reading an image directly, the action reads a bytearray
instead. This again makes it more flexible, so each client can decide
how to handle the byte representation.

Test: existing
Bug: 426529168
Change-Id: I453231f614b5ccada5765bf6ae13da39ec666efa