Add a completion contributor for data sources
Weather data sources can also be in the shape of a pattern, for example:
WEATHER.DAYS.0.IS_AVAILABLE. In this case, the 0 can be an arbitrary
number. We inform the user of that by using a "<days>" or "<hours>"
token in the variants. When autocompleting those lookup elements, we
remove the token and place the cursor where the token was for the user
to specify the days/hours value.

If the user has already specified a days/hours, then we replace the
token with that value in the lookup strings.

Bug: 392849349
Bug: 429175771
Test: updated WFFExpressionCompletionContributorTest
Change-Id: Ic6f6b338bf6718cd640974a35ca99ed2aa066ca4