Revamp BooleanFlag to be drive by text file
We want to move away from an API that allows setting
wrong values. The main issue is creating a BooleanFlag
to guard a feature-in-development, and passing true
instead of enableUpTo(CANARY).

We also want to better control flags for features:
- audit flag differences between 2 releases
- possibly build tooling to ensure the right flags
  are turned on based on go/no-go or launchcal

In order to do this we are changing BooleanFlag to
not accept a default value in the constructor. Instead,
the value is always false by default and can be overriden
via a flag that is tightly controlled.

The file will allow us to control changes via OWNERS,
and to diff between release as the entries are sorted
to ensure minimum changes over time.

With these changes, all implementation of FlagDefault<Boolean>
have been removed, as well as support class (e.g. IdeChannel).

The values are provided to the flags via the override
mechanism of the Flags class. This is probably not
the best place (and it changes the Flag.isOverride
behavior, though we may remove it entirely soon), but
this could change in the near future. This way allowed
us to limit the number of changes necessary to land
this.

Migration:
In order to make sure that the new flag storage is
correct, we compare it to the flag_migration.txt
added previously. A later commit will remove
the migratio file.

Bug: N/A
Test: updated tests
Change-Id: I9ae4644e2018ce709e9344ccc5cef7c92a5b94ba