Fix data source completion
This CL fixes a couple of issues:
1) When autocompleting weather data patterns, like
[WEATHER.DAYS.<days>.IS_AVAILABLE], the token resulting string was a bit
off: [WEATHER.DAYS>.IS_AVAILABLE]. This CL fixes it so that it
becomes the expected: [WEATHER.DAYS..IS_AVAILABLE]

2) When autocompleting after a [, for example: "[WEATHER", the auto
complete was working fine in the tests but not when running manually.
This was because the tests did not wrap the expressions in a watch
face file. This resulted in references not being created in the tests.
However, the references were being created when running manually.
The autocompletion prefix is different when there is a reference, as
it uses the full text of the reference. In this case, we expected the
prefix to be "WEATHER", which is what we got in the tests. However,
when running manually the prefix was the text of the reference:
"[WEATHER".  This led to the autocomplete variants not matching the
prefix and not showing up.

In the tests, we now wrap the expressions in a watch face file to
have the same behaviour as when we run the code manually.

Bug: 392849349
Bug: 429175771
Test: Updated existing to ensure they run in similar conditions to
  manually.
Change-Id: Ibf7cecd0c24d111dcb4cc78bbe6d2a524623f2d6