Remove exclusive-if-local tag for sync benchmarks
This tag should normally apply to tests that run on local machines (i.e.
Mac), but it is being applied to all the OS and resulting in different
performance outcomes.

We still need to add the block-network tag back to Linux and Windows
tests, but this requires tweaking the bazel rules, and will follow up
later.

Bug: N/A
Test: this
Change-Id: Ief47ce29400064ca4515cdcd1428b348a73d980c