Expect duplicate requests when fetching dependency sources
The IntelliJ 2025.2 uses ArtifactView API to resolve auxiliary
dependencies by default as part of 11adacf43d commit and
encapsulated under AuxiliaryArtifactResolverImpl implementation.

However, this causes duplicate requests to fetch sources.jar
when those couldn't be found, applying for both project/buildSrc
dependenciesi. Issue is being tracked on IJPL-199548.

Despite that switching to the old implementation fixes the issue,
for now we are going to just assume that there are going to be
duplicate requests when fetching sources, considering this as
really low impact.

Bug: 429542159
Test: DownloadsAnalyzerTest
Change-Id: If8b4843b52103de627d871f11cd0137e379a1e7b