Add a workaround to avoid test exclusion between targets
The MAC target would exclude the Linux and Windows because these use the
same test_filter, and due to the way test_filter attribute is
defined/used, this ends up in the MAC target only running this test
only.
This shows now on Perfgate since we added the MAC split target because
we no longer have any data for the Linux benchmarks.

bug: 378648219
Test: these
Change-Id: Ifaf9b6e1964edfad373331b7bc2a78735a34d978