Use applicationConfigurable for project view default setting
This change temporarily moves the project view default view setting to a
configurable under Advanced Settings (instead of extending advanced
settings directly) so that this feature is properly hidden when the
studio flag is disabled. We will move this back to advanced settings
when
https://youtrack.jetbrains.com/issue/IJPL-192331/Configuring-visibility-of-custom-Advanced-Settings
is addressed or the setting is ready to be visible for all projects.

Bug: 435683747
Test: tests added
Change-Id: If47c5646d577cc49d6dc72fa96985d49251b8866