Fix adding annotation actions with IntelliJ 2025.2
With IntelliJ 2025.2 there are several annotations that got converted
to ModCommand as part of IDEA-369745 effort. However, now when
specifying a physical annotation the following exception is thrown:
'Annotation attributes must be non-physical PSI'

To avoid it a copy of PsiAnnotation is passed to ModCommand, following
IntelliJ practices 7d57941ece.

Bug: 436515426
Test: ReplacePreviewAnnotationFixTest
Change-Id: Ib09479d6ee456adc9f3b12db54167a0b2a58f1b5