Move Git4Idea dep closer to where it is used
Git4Idea does not quite belong in studio-sdk-all-modules because it is
not a v2 module. However, we still need it in all Android plugin test
targets because the Android plugin descriptor fails to load without
it (due to the dependency on intellij.settingsSync.core which itself
depends on Git4Idea).

Bug: not filed
Test: existing
Change-Id: I2b5650beb3e1e80e90670807e6edb0e3ef6bf087