Temporary ignore assertion for safe delete test when using databinding
With IntelliJ 2025.2, safe delete was migrated to the new conflict
dialog as part commit 0fce75621 for IDEA-360408 effort. However,
this allowed to spot a bug in studio-main when safe deleting at the
same time databinding layout resource and file class importing it,
resulting on wrongly identify imports as usages despite that entire
file will be deleted.

For now, the assertion of both files being deleted will be ignored,
to unblock merge, allowing to address this separately.

Bug: 437036798
Test: This
Change-Id: I0c29b5fa0e7f731b7d27d535cd406f1860215753