Add support for string resource resolution in template
... parameters

If a template parameter's expression is a simple string, it can
reference a string resource. However, if it doesn't reference any string
resource it will be considered as a simple string instead.

The reference can be surrounded by quotes.

If in an expression that's not a simple string, it cannot reference a
string resource.

Bug: 437048451
Test: TemplateParameterStringReferenceTest
Change-Id: Ia2e170e5d6e044e7805521ee16d4b6d66d5c863a