Ignore digest not found artifact if build get failed
In some cases, bep output only return exit code not 0 event but no
action. In this case, we cannot find failed target but digest map should
not throw illegal state as the build faile and it's possible artifact
does not exist. So pass both exit code and target with error to
digestMap to avoid unexpected exception is thrown.

Bug: b/436766821
Test: existing tests + manual tests

Change-Id: Ideee97dc3f34b54424a6dbfc7ca24d575b3219de