Show the deprecation notification using invokeLater
When the notification is immediately expired in the test, the variable
ends up null since it is not assigned and leads to the expire part not
running. This lets the notification get assigned to the variable and
leads to the expire part running in tests.

Presubmit-runs-per-test: //tools/adt/idea/android:intellij.android.core.tests_tests@200
Fixes: 437253473
Test: Existing
Change-Id: Ibdcaa02da42f837350ed49855915cb31d12c5216